refactor separate hook useSyncURLSlug
