+++ frontend/src/metabase/dashboard/components/DashboardTabs/useDashboardTabs.ts
@@ -1,56 +1,32 @@
-import { useEffect } from "react";
-import { useMount, usePrevious } from "react-use";
+import { useMount } from "react-use";
 import { t } from "ttag";
 import type { UniqueIdentifier } from "@dnd-kit/core";
 
 import { useDispatch, useSelector } from "metabase/lib/redux";
 import {
-  createNewTab as createNewTabAction,
+  createNewTab,
   renameTab,
   deleteTab as deleteTabAction,
   initTabs,
-  selectTab as selectTabAction,
+  selectTab,
   undoDeleteTab,
   moveTab as moveTabAction,
 } from "metabase/dashboard/actions";
 import { SelectedTabId } from "metabase-types/store";
-import { getDashboardId, getSelectedTabId } from "metabase/dashboard/selectors";
+import { getSelectedTabId, getTabs } from "metabase/dashboard/selectors";
 import { addUndo } from "metabase/redux/undo";
-import { getSlug, useUpdateURLSlug } from "./utils";
+
+import { useSyncURLSlug } from "./useSyncURLSlug";
 
 let tabDeletionId = 1;
 
 export function useDashboardTabs({ slug }: { slug: string | undefined }) {
   const dispatch = useDispatch();
-  const dashboardId = useSelector(getDashboardId);
-  const tabs = useSelector(state =>
-    dashboardId
-      ? state.dashboard.dashboards[dashboardId].ordered_tabs?.filter(
-          tab => !tab.isRemoved,
-        ) ?? []
-      : [],
-  );
+  const tabs = useSelector(getTabs);
   const selectedTabId = useSelector(getSelectedTabId);
-  const prevSelectedTabId = usePrevious(selectedTabId);
-  const { updateURLSlug } = useUpdateURLSlug();
 
+  useSyncURLSlug();
   useMount(() => dispatch(initTabs({ slug })));
-
-  useEffect(() => {
-    if (selectedTabId !== prevSelectedTabId) {
-      updateURLSlug(
-        getSlug({
-          tabId: selectedTabId,
-          name: tabs.find(t => t.id === selectedTabId)?.name,
-        }),
-      );
-    }
-  }, [selectedTabId, prevSelectedTabId, tabs, updateURLSlug]);
-
-  const createNewTab = () => {
-    dispatch(createNewTabAction());
-    // updateURLSlug("");
-  };
 
   const deleteTab = (tabId: SelectedTabId) => {
     const tabName = tabs.find(({ id }) => id === tabId)?.name;
@@ -78,23 +54,14 @@ export function useDashboardTabs({ slug }: { slug: string | undefined }) {
       }),
     );
 
-  const selectTab = (tabId: SelectedTabId) => {
-    dispatch(selectTabAction({ tabId }));
-
-    // const name = tabs.find(t => t.id === tabId)?.name;
-    // if (name) {
-    //   updateURLSlug(getSlug({ tabId, name }));
-    // }
-  };
-
   return {
     tabs,
     selectedTabId,
-    createNewTab,
+    createNewTab: () => dispatch(createNewTab()),
     deleteTab,
     renameTab: (tabId: SelectedTabId, name: string) =>
       dispatch(renameTab({ tabId, name })),
-    selectTab,
+    selectTab: (tabId: SelectedTabId) => dispatch(selectTab({ tabId })),
     moveTab,
   };
 }
