+++ frontend/src/metabase/dashboard/selectors/index.ts
@@ -1,2 +1,2 @@
 export * from "./selectors";
-export * from "./selectors_typed";
+export * from "./selectors-typed";
