+++ frontend/src/metabase/dashboard/selectors/selectors.js
@@ -17,7 +17,7 @@ import { getEmbedOptions, getIsEmbedded } from "metabase/selectors/embed";
 import Question from "metabase-lib/Question";
 
 import { isVirtualDashCard } from "../utils";
-import { getDashboardId } from "./selectors_typed";
+import { getDashboardId } from "./selectors-typed";
 
 export const getIsEditing = state => !!state.dashboard.isEditing;
 export const getDashboardBeforeEditing = state => state.dashboard.isEditing;
